fix product CRUD change product-category nullable = true change product_price from BigDecimal to float

+++ spring-backend/src/main/java/sg/nus/iss/spring/backend/interfacemethods/ProductService.java
@@ -2,11 +2,10 @@ package sg.nus.iss.spring.backend.interfacemethods;
 
 import sg.nus.iss.spring.backend.model.Product;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
-    List<Product>  list(Integer categoryId, BigDecimal minPrice, BigDecimal maxPrice, String keywords);
+    List<Product>  list(Integer categoryId, Float minPrice, Float maxPrice, String keywords);
     Product findProductById(Integer productId);
     Product createProduct(Product product);
     Product editProduct(Product product);
